Split up the metrics orchestration code
Divided between what will be re-usable with the fetch metrics helpers and
what is specific to the device events.  The fetch metrics will not have
the `LocalInstanceGroup` data available like the device metrics.

Bug: 491965176
