+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -113,7 +113,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/instances:cvd_persistent_data",
         "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/utils",
-        "//cuttlefish/host/libs/metrics:metrics_orchestration",
+        "//cuttlefish/host/libs/metrics:device_metrics_orchestration",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/result",
