+++ base/cvd/cuttlefish/host/libs/metrics/device_metrics_orchestration.h
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include "cuttlefish/host/commands/cvd/instances/local_instance_group.h"
+
+namespace cuttlefish {
+
+void GatherVmInstantiationMetrics(const LocalInstanceGroup& instance_group);
+
+void GatherVmStartMetrics(const LocalInstanceGroup& instance_group);
+
+void GatherVmBootCompleteMetrics(const LocalInstanceGroup& instance_group);
+
+void GatherVmBootFailedMetrics(const LocalInstanceGroup& instance_group);
+
+void GatherVmStopMetrics(const LocalInstanceGroup& instance_group);
+
+}  // namespace cuttlefish
